pyproject.toml
[project]
name = "simple-chatagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "chromadb>=0.4.24",
    "deep-translator>=1.11.4",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-huggingface>=0.3.0",
    "openai-agents>=0.0.19",
    "pinecone>=7.2.0",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.0",
    "pypdf2>=3.0.1",
    "sentence-transformers>=4.1.0",
]
